# openapi/sunbloom-api.json
{
    "openapi": "3.1.1",
    "info": {
        "title": "SunBloom.Api | v1",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:5078/"
        }
    ],
    "paths": {
        "/api/v1/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Create an account.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpValidationProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Exchange credentials for tokens.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpValidationProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Rotate a refresh token.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RefreshRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/problem+json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HttpValidationProblemDetails"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Revoke the current token family.",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RefreshRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/api/v1/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "The authenticated user.",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/skills/tree": {
            "get": {
                "tags": [
                    "Skills"
                ],
                "summary": "The skill containment tree, optionally rooted at one skill.",
                "parameters": [
                    {
                        "name": "root",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SkillTreeNode"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/skills/{slug}": {
            "get": {
                "tags": [
                    "Skills"
                ],
                "summary": "One skill with its prerequisites, unlocks, and related skills.",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SkillDetail"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "AuthResponse": {
                "required": [
                    "accessToken",
                    "refreshToken",
                    "expiresInSeconds",
                    "user"
                ],
                "type": "object",
                "properties": {
                    "accessToken": {
                        "type": "string"
                    },
                    "refreshToken": {
                        "type": "string"
                    },
                    "expiresInSeconds": {
                        "pattern": "^-?(?:0|[1-9]\\d*)$",
                        "type": [
                            "integer",
                            "string"
                        ],
                        "format": "int32"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserResponse"
                    }
                }
            },
            "HttpValidationProblemDetails": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "title": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "status": {
                        "pattern": "^-?(?:0|[1-9]\\d*)$",
                        "type": [
                            "null",
                            "integer",
                            "string"
                        ],
                        "format": "int32"
                    },
                    "detail": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "instance": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "LoginRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                }
            },
            "RefreshRequest": {
                "required": [
                    "refreshToken"
                ],
                "type": "object",
                "properties": {
                    "refreshToken": {
                        "type": "string"
                    }
                }
            },
            "RegisterRequest": {
                "required": [
                    "email",
                    "password",
                    "displayName",
                    "timeZone"
                ],
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "displayName": {
                        "type": "string"
                    },
                    "timeZone": {
                        "type": [
                            "null",
                            "string"
                        ]
                    }
                }
            },
            "SkillDetail": {
                "required": [
                    "id",
                    "slug",
                    "name",
                    "description",
                    "kind",
                    "parentSkillId",
                    "prerequisites",
                    "unlocks",
                    "related"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "kind": {
                        "type": "string"
                    },
                    "parentSkillId": {
                        "type": [
                            "null",
                            "string"
                        ],
                        "format": "uuid"
                    },
                    "prerequisites": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SkillSummary"
                        }
                    },
                    "unlocks": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SkillSummary"
                        }
                    },
                    "related": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SkillSummary"
                        }
                    }
                }
            },
            "SkillSummary": {
                "required": [
                    "id",
                    "slug",
                    "name",
                    "kind"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "kind": {
                        "type": "string"
                    }
                }
            },
            "SkillTreeNode": {
                "required": [
                    "id",
                    "slug",
                    "name",
                    "description",
                    "kind",
                    "children"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": [
                            "null",
                            "string"
                        ]
                    },
                    "kind": {
                        "type": "string"
                    },
                    "children": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SkillTreeNode"
                        }
                    }
                }
            },
            "UserResponse": {
                "required": [
                    "id",
                    "email",
                    "displayName",
                    "timeZone"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "email": {
                        "type": "string"
                    },
                    "displayName": {
                        "type": "string"
                    },
                    "timeZone": {
                        "type": "string"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Auth"
        },
        {
            "name": "Skills"
        }
    ]
}